Don't require package-lock.json to be created
A bunch of services set the package-lock not to be created.  The script would
error on the last step in that case.  That's not terrible, but code tastes
better without crashes.

Co-author: @erikap

+++ scripts/install/run.sh
@@ -12,4 +12,7 @@ cd /usr/src/app
 NODE_ENV=development ./install-dependencies.sh
 
 docker-rsync /usr/src/app/app/node_modules/ /data/service/node_modules
-cp /usr/src/app/app/package-lock.json /data/service/package-lock.json
+if [ -f /usr/src/app/app/package-lock.json ]
+then
+   cp /usr/src/app/app/package-lock.json /data/service/package-lock.json
+fi
